Extract common UUID validation to its own VO
