+++ src/modules/Shared/Domain/ValueObject/ValueObject.ts
