+++ src/modules/User/Domain/ValueObject/UserUsername.ts
@@ -1,4 +1,4 @@
-import { ValueObject } from '~/src/modules/Shared/Domain/ValueObject'
+import { ValueObject } from '~/src/modules/Shared/Domain/ValueObject/ValueObject'
 import { UserDomainException } from '~/src/modules/User/Domain/UserDomainException'
 
 export class UserUsername extends ValueObject<string> {
